Only skip DEBUG_STREAM when it is not default_stream
There is a bug in serial_available() that will make the system ignore any incoming bytes from default_stream.
With UART as default_stream, the system will not be able to communicate with a sender software on a host machine.

To repro:
1. enable DEBUG_STREAM (mapped to default_stream)
2. register multiple serial streams
3. build and flash
4. connect system to a sender, and verify the two can communicate
5. send data via any stream that is not default_stream (i.e. not UART)
6. sender can no longer communicate with the system

This diff fixes the bug by skipping DEBUG_STREAM only when it is not default_stream.